add /scan command, always uses all engines

+++ src/label.cpp
@@ -17,7 +17,6 @@
  * limitations under the License.
  *
  ************************************************************************************/
-#include <typeinfo>
 #include <dpp/dpp.h>
 #include <beholder/config.h>
 #include <beholder/beholder.h>
@@ -28,7 +27,7 @@
 
 namespace label {
 
-	bool scan(bool &flattened, const std::string& hash, std::string& file_content, const dpp::attachment& attach, dpp::cluster& bot, const dpp::message_create_t ev, int pass) {
+	bool scan(bool &flattened, const std::string& hash, std::string& file_content, const dpp::attachment& attach, dpp::cluster& bot, const dpp::message_create_t ev, int pass, bool delete_message) {
 		db::resultset settings = db::query("SELECT guild_id FROM guild_config WHERE guild_id = ? AND objects_this_month <= object_limit", { ev.msg.guild_id });
 		if (settings.empty()) {
 			/* Not configured, or out of quota */
@@ -101,7 +100,11 @@ namespace label {
 				query += ")";
 				auto check = db::query(query, params);
 				if (!check.empty()) {
-					delete_message_and_warn(hash, file_content, bot, ev, attach, human, false);
+					if (delete_message) {
+						delete_message_and_warn(hash, file_content, bot, ev, attach, human, false);
+					} else {
+						throw std::runtime_error(human);
+					}
 					INCREMENT_STATISTIC("images_label", ev.msg.guild_id);
 					return true;
